ai-coding-course-claude/lessons/lesson-05-reasoning-models/src/claude-5/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cli-transcript-analyzer"
version = "0.1.0"
description = "CLI application for transcript analysis using Gemini AI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.0",
    "pydantic>=2.0.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
main = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]